Rewrote HashMap and added auto scroll to console

+++ Engine/Source/Oniun/RHI/Base/IImGuiLayer.cpp
@@ -83,7 +83,7 @@ void IImGuiLayer::SetFont(const StringView& font, uint64 fontSize)
 
     if (!setDefaults)
     {
-        if (!io.Fonts->AddFontFromFileTTF(*font, fontSize))
+        if (!io.Fonts->AddFontFromFileTTF(*font, (float)fontSize))
             setDefaults = true;
     }
